requirements.txt
openai
anthropic
httpx
python-dotenv
pandas
dashscope